requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
pydantic
ddgs
duckduckgo-search
arxiv
praw
flask
Flask-CORS